+++ paddle.h
@@ -11,13 +11,13 @@ public:
   Paddle(bool is_left);
 
   bool is_left() const { return left_side_; }
-  void move(const double dt, const bool direction);
+  void move(const float dt, const bool direction);
   void reset();
   void increase_speed();
 
-  inline double get_x() const { return x_pos_; }
-  inline double get_y() const { return y_pos_; }
-  inline double get_speed() const { return speed_; }
+  inline float get_x() const { return x_pos_; }
+  inline float get_y() const { return y_pos_; }
+  inline float get_speed() const { return speed_; }
 
   bool check_collision(const Ball& b) const;
 
@@ -26,12 +26,12 @@ public:
   friend class Bot;
 
 private:
-  void set_y(const double y);
+  void set_y(const float y);
 
   bool left_side_;  // left = true, right = false
-  double y_pos_;    // Position of centre of paddle
-  const double x_pos_;
-  double speed_;
+  float y_pos_;    // Position of centre of paddle
+  const float x_pos_;
+  float speed_;
 };
 
 #endif // PADDLE_H
